JS: Checking row count before deletion

+++ resources/js/toolbar/toolbar.js
@@ -7,13 +7,19 @@ export function deleteSelectedRowsFromToolbar(table_id, model, id_column, succes
     //! Most likely due to variable scoping, so I just changed it to be a string
     // Get selected table rows
     var selectedRows = $('#' + table_id).bootstrapTable('getSelections');
-    // console.log("selectedRows: ", selectedRows);
+    console.log("selectedRows: ", selectedRows);
 
-    // Extract IDs
-    var ids = selectedRows.map(obj => obj._data.id);
+    if (selectedRows.length === 0) {
+        console.log("Nothing selected")
+        alert("Please select row(s) to be deleted.\nYou can use Ctrl and Shift to select multiple rows");
+    }
+    else {
+        // Extract IDs
+        var ids = selectedRows.map(obj => obj._data.id);
 
-    if (confirm('Are you sure you want to delete ' + selectedRows.length + ' selected row(s)?\n\nThis will also delete the corresponding entries from BOMs, storage locations and stock history.')) {
-        deleteSelectedRows(ids, model, id_column, successCallback);
+        if (confirm('Are you sure you want to delete ' + selectedRows.length + ' selected row(s)?\n\nThis will also delete the corresponding entries from BOMs, storage locations and stock history.')) {
+            deleteSelectedRows(ids, model, id_column, successCallback);
+        }
     }
 }
